refactor(PageTitle): move component to its own directory
