+++ src/pages/analysis/analysis/index.tsx
@@ -1,4 +1,4 @@
-import PageTitle from '~components/titles/PageTitle'
+import PageTitle from '~/components/titles/PageTitle/PageTitle'
 
 export default function AnalysisReport() {
     return (
