+++ src/pages/analysis/quantity/index.tsx
@@ -1,4 +1,4 @@
-import PageTitle from '~components/titles/PageTitle'
+import PageTitle from '~/components/titles/PageTitle/PageTitle'
 
 export default function Quantity() {
     return (
