+++ src/pages/parametros/categories/index.tsx
@@ -3,7 +3,7 @@ import { CircularProgress } from '@mui/material'
 
 import AllCategoriesTable from '~components/settings/categories/AllCategoriesTable'
 import EditToolbar from '~components/table/headers/toolbar'
-import PageTitle from '~components/titles/PageTitle'
+import PageTitle from '~/components/titles/PageTitle/PageTitle'
 import { useGetAllCategoriesQuery } from '~redux/api/bca-backend/parametros/categoriesSlice'
 import CategoriesDrawer from '~components/drawers/Settings/Categories/CategoriesDrawer'
