+++ src/pages/parametros/items/detail/index.tsx
@@ -3,7 +3,7 @@ import { useLocation } from 'react-router-dom'
 import { Box, CircularProgress } from '@mui/material'
 
 import { useGetOneRubroQuery } from '~redux/api/bca-backend/parametros/rubrosSlice'
-import PageTitle from '~components/titles/PageTitle'
+import PageTitle from '~/components/titles/PageTitle/PageTitle'
 import EditToolbar from '~components/table/headers/toolbar'
 import AllRubrosMaterialsTable from '~/components/settings/rubros/AllRubrosMaterialsTable/AllRubrosMaterialsTable'
 import RubroMaterialsDrawer from '~components/drawers/Settings/RubroMaterial/RubroMaterialsDrawer'
