+++ src/pages/parametros/items/index.tsx
@@ -3,7 +3,7 @@ import { CircularProgress } from '@mui/material'
 
 import AllRubrosTable from '~components/settings/rubros/AllRubrosTable'
 import EditToolbar from '~components/table/headers/toolbar'
-import PageTitle from '~components/titles/PageTitle'
+import PageTitle from '~/components/titles/PageTitle/PageTitle'
 import { useGetAllRubrosQuery } from '~redux/api/bca-backend/parametros/rubrosSlice'
 
 export default function Items() {
