+++ src/pages/parametros/materials/index.tsx
@@ -3,7 +3,7 @@ import { CircularProgress } from '@mui/material'
 
 import AllMaterialsTable from '~components/settings/materials/AllMaterialsTable'
 import EditToolbar from '~components/table/headers/toolbar'
-import PageTitle from '~components/titles/PageTitle'
+import PageTitle from '~/components/titles/PageTitle/PageTitle'
 import { useGetAllMaterialsQuery } from '~redux/api/bca-backend/parametros/materialsSlice'
 import MaterialsDrawer from '~components/drawers/Settings/Materials/MaterialsDrawer'
