+++ src/pages/transacciones/factura/index.tsx
@@ -3,7 +3,7 @@ import { CircularProgress } from '@mui/material'
 
 import AllInvoicesTable from '~components/parameters/invoices/AllInvoicesTable'
 import EditToolbar from '~components/table/headers/toolbar'
-import PageTitle from '~components/titles/PageTitle'
+import PageTitle from '~/components/titles/PageTitle/PageTitle'
 import { useGetAllInvoicesQuery } from '~redux/api/bca-backend/transacciones/invoiceSlice'
 
 export default function Factura() {
